requirements.txt
fastapi[all]
sqlmodel
alembic
fastapi-cache2[redis]
sqladmin
python-jose[cryptography]
asyncpg
gunicorn
passlib